Fix signed/unsigned warning in vtkCSGGrid.C
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@26583 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/visit_vtk/full/vtkCSGGrid.C
@@ -2795,8 +2795,8 @@ vtkCSGGrid::ExtractRegionBounds(int specificZone, int &nRegionBounds,
                << regionBounds2[i*11+10] << endl;
     }
     zoneMap2 = new int[numBoundaries];
-    for (size_t i = 0; i < nRegionBounds; i++)
-        zoneMap2[(int)fabs(regionBounds2[i*11])] = (int)i;
+    for (int i = 0; i < nRegionBounds; i++)
+        zoneMap2[(int)fabs(regionBounds2[i*11])] = i;
 
     int nRegionBounds3 = 0;
     double *regionBounds3 = new double[nRegionBounds*10];
